Extract NSE specific code out of the UNNotificationContent extension
